デザインを一掃した
