--- frontend/src/components/Weight/styles/WeightChart.css
@@ -1,62 +1,0 @@
-.weight-chart-container {
-    width: 100%;
-    max-width: 800px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: rgba(162, 217, 255, 0.1);;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .weight-chart-container h2 {
-    text-align: center;
-    color: #0366d6;
-    margin-bottom: 20px;
-  }
-  
-  .weight-chart-wrapper {
-    position: relative;
-    height: 400px;
-  }
-
-.weight-chart-title {
-  font-size: 24px; /* タイトルのフォントサイズを調整 */
-}
-
-.weight-chart-controls {
-  margin-bottom: 20px;
-  display: flex;
-  align-items: center;
-}
-
-.weight-chart-controls label {
-  margin-right: 10px;
-  font-size: 18px; /* ラベルのフォントサイズを調整 */
-}
-
-.weight-chart-controls select {
-  font-size: 15px; /* 選択欄のフォントサイズを調整 */
-  padding: 5px;
-  background-color: #fff; /* 薄い青の背景色 */
-  border: 1px solid #b3d7ff; /* 薄い青の境界線 */
-  border-radius: 5px; /* 角を丸くする */
-  color: #0366d6; /* テキストの色 */
-}
-
-.weight-x-axis-title,
-.weight-y-axis-title {
-  font-size: 14px; /* 軸タイトルのフォントサイズを調整 */
-}
-
-.weight-x-axis-ticks,
-.weight-y-axis-ticks {
-  font-size: 12px; /* 軸目盛りのフォントサイズを調整 */
-}
-
-.weight-legend-labels {
-  font-size: 14px; /* 凡例ラベルのフォントサイズを調整 */
-}
-
-.weight-tooltip {
-  font-size: 12px; /* ツールチップのフォントサイズを調整 */
-}
